fix: 말풍선 UI 변경
